Get data from exercises or projects for one student

Get data from exercises or projects for one student

get_data(
  collection,
  query,
  fields = "{}",
  url = getOption("learnitr.lrs_url"),
  count = FALSE
)

get_profile(email, login = NULL)

get_all_data(email, icourse, module = NULL)

get_github_data(email)

get_grid_data(email, icourse, module = NULL)

get_grade_data(email, icourse)

get_grade_dir(root_dir = "/data1/grades", acad_year = NULL)

Arguments

collection

The collection in the MongoDB database

query

The JSON query on the MongoDB database

fields

The fields to return

url

The url to access the MongoDB database

count

Do we count items instead of getting them?

email

The email of the student

login

The GitHub login of the student as alternate way to get its profile (when it is not NULL)

icourse

The course identifier

module

The module, or NULL for exercises in all modules

root_dir

The root directory where the grades are stored.

acad_year

The academic year to use, or NULL to use the latest one.

Value

A data frame or a list with the data